requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.8.1
cohere==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
google-generativeai==0.4.1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6